+++ kopy-compiler/main/kotlin/com/javiersc/kotlin/kopy/compiler/fir/FirKopyExtension.kt
@@ -1,7 +1,6 @@
 package com.javiersc.kotlin.kopy.compiler.fir
 
 import com.javiersc.kotlin.kopy.compiler.fir.checker.FirKopyCheckerExtension
-import com.javiersc.kotlin.kopy.compiler.fir.generation.FirCopySupertypeGenerationExtension
 import com.javiersc.kotlin.kopy.compiler.fir.generation.FirKopyAssignExpressionAltererExtension
 import com.javiersc.kotlin.kopy.compiler.fir.generation.FirKopyDeclarationGenerationExtension
 import org.jetbrains.kotlin.GeneratedDeclarationKey
@@ -23,7 +22,6 @@ internal class FirKopyExtension(
 
     private fun ExtensionRegistrarContext.registerGenerators() {
         +::FirKopyAssignExpressionAltererExtension
-        +::FirCopySupertypeGenerationExtension
         +::FirKopyDeclarationGenerationExtension
     }
 }