+++ kopy-compiler/main/kotlin/com/javiersc/kotlin/kopy/compiler/ir/IrKopyGenerationExtension.kt
@@ -1,7 +1,7 @@
 package com.javiersc.kotlin.kopy.compiler.ir
 
 import com.javiersc.kotlin.kopy.compiler.ir.transformers.IrAtomicPropertyTransformer
-import com.javiersc.kotlin.kopy.compiler.ir.transformers.IrInitKopyableFunctionTransformer
+import com.javiersc.kotlin.kopy.compiler.ir.transformers.IrFunctionsTransformer
 import com.javiersc.kotlin.kopy.compiler.ir.transformers.IrSetOrUpdateCallTransformer
 import com.javiersc.kotlin.kopy.compiler.ir.transformers.IrUpdateEachCallTransformer
 import org.jetbrains.kotlin.backend.common.extensions.IrGenerationExtension
@@ -22,7 +22,7 @@ internal class IrKopyGenerationExtension(
     @JvmName("generate2")
     private fun IrModuleFragment.generate(pluginContext: IrPluginContext) {
         transform(IrAtomicPropertyTransformer(this, pluginContext))
-        transform(IrInitKopyableFunctionTransformer(this, pluginContext))
+        transform(IrFunctionsTransformer(this, pluginContext))
         transform(IrUpdateEachCallTransformer(this, pluginContext))
         transform(IrSetOrUpdateCallTransformer(this, pluginContext))